pyproject.toml
[tool.poetry]
name = "innotter"
version = "0.1.0"
description = ""
authors = ["Granit319 <hleb.petrutsik@innowise-group.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg-binary = "^3.0.15"
isort = "^5.10.1"
black = "^22.6.0"
bandit = "^1.7.4"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
